requirements.txt
numpy==1.18.4
pandas==1.0.3
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
